retrieval/requirements.txt
accelerate==1.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
bitsandbytes==0.46.1
cbor==1.0.0
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
datasets==4.0.0
deepspeed==0.17.4
dill==0.3.8
einops==0.8.1
FlagEmbedding==1.3.5
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.5
hjson==3.1.0
huggingface-hub==0.34.3
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir-datasets==0.5.11
ir-measures==0.3.7
jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
lz4==4.4.4
markupsafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.0
pandas==2.0.1
peft==0.10.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
pytrec-eval-terrier==0.5.7
pytz==2025.2
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.49.0
trec-car-tools==2.6
triton==3.3.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
xformers==0.0.31.post1
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9